requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
beautifulsoup4==4.12.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cssselect==1.3.0
exceptiongroup==1.3.1
fastapi==0.115.5
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.20.2
h11==0.16.0
httptools==0.7.1
idna==3.11
jieba3k==0.35.1
joblib==1.5.3
lxml==5.3.0
lxml_html_clean==0.1.1
newspaper3k==0.2.8
nltk==3.9.1
pillow==12.1.0
pydantic==2.10.3
pydantic-settings==2.2.1
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.3
requests-file==3.0.1
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.8.1
starlette==0.41.3
tinysegmenter==0.3
tldextract==5.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.32.1
watchfiles==1.1.1
websockets==15.0.1